Pulled the remaining configutation code for the student modal and put it into a hook for isolation.
