+++ client/src/pages/Dashboard/components/StudentModal/index.js
@@ -1,12 +1,7 @@
-import { useMemo } from "react";
-
 import {
 	Box,
 	Columns
 } from "react-bulma-components";
-
-import { getDashboardAction as gda, useEditStudent, useClassroom, useDashboardContext } from "pages/Dashboard/store";
-import { EDIT_STUDENT } from "pages/Dashboard/store/actionsNames";
 
 import {
 	ModalBox,
@@ -14,40 +9,31 @@ import {
 
 import {
 	PanelTabs,
-	usePanels
 } from "./panels";
 
-import {useWindowDimensions} from "utils/windowWidth"
+import { useStudentModalConfig } from "./utils";
 
-import "./index.sass" 
+import "./index.sass" ;
 
 const { Column } = Columns;
 
 const StudentModal = () => {
 
 	// We pull in the student to edit from the dashboard state
-    const [{ editStudent: editStudentId }, dispatch] = useDashboardContext();
-
-    const isViewing = editStudentId !== false;
-
-    const { _id: roomId } = useClassroom();
-
-    const editStudent = useEditStudent();
-
-    const clearEditStudent = () => dispatch(gda(EDIT_STUDENT, false));
-
-    const { width } = useWindowDimensions();
-	const windowBreakPoint = 1025;
-
-	const panelConfig = usePanels( roomId, editStudent );
-
-	const activePanels = useMemo(() => new Map( [...panelConfig.panels].filter(([key]) => width >= windowBreakPoint || key === panelConfig.activePanel) ));
+	const {
+		student,
+		isViewing,
+		panelConfig,
+		activePanels,
+		clearEditStudent
+	} = useStudentModalConfig();
 
     return (
-		<ModalBox fullScreen={editStudent._id} onClose={clearEditStudent} show={isViewing}>
-
+		<ModalBox fullScreen={student._id} onClose={clearEditStudent} show={isViewing}>
+			{/* Dispay panel tabs */}
 			{panelConfig.panels.size > 1 && <Box className="is-tabs"><PanelTabs {...panelConfig} /></Box>}
 
+			{/* Display each active panel */}
 			{[...activePanels].map(([key,{Panel}]) => (
 				<Column
 					key={key}
